+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/result/CoreResponseProvider.kt
@@ -1,5 +1,0 @@
-package com.mapbox.search.result
-
-internal interface CoreResponseProvider {
-    val originalSearchResult: OriginalSearchResult
-}